+++ src/refactoring/components/admin/coupon/CouponBox.tsx
@@ -1,4 +1,4 @@
-import { Coupon } from '../../../types'
+import { Coupon } from '../../../../types'
 
 interface Props {
   coupon: Coupon
