+++ src/refactoring/components/admin/product/CreateProductFormBox.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../../types'
+import { Product } from '../../../../types'
 
 interface Props {
   newProduct: Omit<Product, 'id'>
