+++ src/refactoring/components/admin/product/ProductManagement.tsx
@@ -1,4 +1,4 @@
-import { Discount, Product } from '../../../types'
+import { Discount, Product } from '../../../../types'
 import { CreateProductFormBox } from './CreateProductFormBox'
 import { ProductList } from './ProductList'
